+++ Covenant/Core/CovenantHubService.cs
@@ -267,6 +267,11 @@ namespace Covenant.Core
             return _connection.InvokeAsync<CapturedTicketCredential>("CreateTicketCredential", credential);
         }
 
+        public Task<Theme> CreateTheme(Theme theme)
+        {
+            return _connection.InvokeAsync<Theme>("CreateTheme", theme);
+        }
+
         public Task<CovenantUser> CreateUser(CovenantUserLogin login)
         {
             return _connection.InvokeAsync<CovenantUser>("CreateUser", login);
@@ -350,6 +355,11 @@ namespace Covenant.Core
         public Task DeleteReferenceSourceLibrary(int id)
         {
             return _connection.InvokeAsync("DeleteReferenceSourceLibrary", id);
+        }
+
+        public Task DeleteTheme(int id)
+        {
+            return _connection.InvokeAsync("DeleteTheme", id);
         }
 
         public Task DeleteUser(string userId)
@@ -507,9 +517,19 @@ namespace Covenant.Core
             return _connection.InvokeAsync<CapturedTicketCredential>("EditTicketCredential", credential);
         }
 
-        public Task<CovenantUser> EditUser(CovenantUser currentUser, CovenantUserLogin user)
+        public Task<Theme> EditTheme(Theme theme)
         {
-            return _connection.InvokeAsync<CovenantUser>("EditUser", currentUser, user);
+            return _connection.InvokeAsync<Theme>("EditTheme", theme);
+        }
+
+        public Task<CovenantUser> EditUser(CovenantUser currentUser)
+        {
+            return _connection.InvokeAsync<CovenantUser>("EditUser", currentUser);
+        }
+
+        public Task<CovenantUser> EditUserPassword(CovenantUser currentUser, CovenantUserLogin user)
+        {
+            return _connection.InvokeAsync<CovenantUser>("EditUserPassword", currentUser, user);
         }
 
         public Task<WmicLauncher> EditWmicLauncher(WmicLauncher launcher)
@@ -1090,6 +1110,16 @@ namespace Covenant.Core
         public Task<IEnumerable<TargetIndicator>> GetTargetIndicators()
         {
             return _connection.InvokeAsync<IEnumerable<TargetIndicator>>("GetTargetIndicators");
+        }
+
+        public Task<Theme> GetTheme(int id)
+        {
+            return _connection.InvokeAsync<Theme>("GetTheme", id);
+        }
+
+        public Task<IEnumerable<Theme>> GetThemes()
+        {
+            return _connection.InvokeAsync<IEnumerable<Theme>>("GetThemes");
         }
 
         public Task<CapturedTicketCredential> GetTicketCredential(int credentialId)